🎨 Moved all the relationships into one module.
